Refactor: separadas las funciones de API en archivos especificos
