+++ src/api/convocatorias.api.ts
@@ -1,0 +1,38 @@
+import axios from "axios";
+
+export const postConvocatoria = (formData: Object) => {
+    axios
+        .post("http://localhost:3000/convocatoria", formData, {
+            headers: { "Content-Type": "multipart/form-data" }
+        })
+        .then(function (response) {
+            console.log(response);
+        })
+        .catch(function (error) {
+            console.error(error);
+        });
+};
+
+export const getConvocatorias = async () => {
+    const response = await axios.get("http://localhost:3000/convocatoria");
+    return response.data;
+};
+
+export const patchFechaConvocatoria = async (id: string, fechaFin: Date) => {
+    console.log({ fechaFin })
+    await axios
+        .patch(`http://localhost:3000/convocatoria/${id}/fecha-fin`, { fechaFin })
+        .then(function (response) {
+            console.log(response);
+        })
+};
+
+export const deleteConvocatoria = async (id: string): Promise<void> => {
+    try {
+        await axios.delete(`http://localhost:3000/convocatoria/${id}`);
+        console.log("Convocatoria eliminada correctamente");
+    } catch (error) {
+        console.error("Error al eliminar la convocatoria", error);
+        throw error;
+    }
+};
