+++ src/api/usuarios.api.ts
@@ -1,68 +1,4 @@
 import axios from "axios";
-import { NuevoFormatoValues } from "../pages/FormNuevaConvocatoria/schemas/nuevoFormatoSchema";
-
-export const postFormato = (formato: NuevoFormatoValues) => {
-    axios
-        .post("http://localhost:3000/formato", formato)
-        .then(function (response) {
-            console.log(response);
-        })
-        .catch(function (error) {
-            console.error(error);
-        });
-}
-
-export const getFormatos = async () => {
-    const response = await axios.get("http://localhost:3000/formato")
-    return response.data
-}
-
-export const getFormatoById = async (id: string) => {
-    const response = await axios.get(`http://localhost:3000/formato/${id}`)
-    return response.data
-}
-
-export const getFormatoByNombre = async (nombre: string) => {
-    const response = await axios.get(`http://localhost:3000/formato/nombre/${nombre}`);
-    return response.data;
-}
-
-export const postConvocatoria = (formData: Object) => {
-    axios
-        .post("http://localhost:3000/convocatoria", formData, {
-            headers: { "Content-Type": "multipart/form-data" }
-        })
-        .then(function (response) {
-            console.log(response);
-        })
-        .catch(function (error) {
-            console.error(error);
-        });
-};
-
-export const getConvocatorias = async () => {
-    const response = await axios.get("http://localhost:3000/convocatoria");
-    return response.data;
-};
-
-export const patchFechaConvocatoria = async (id: string, fechaFin: Date) => {
-    console.log({ fechaFin })
-    await axios
-        .patch(`http://localhost:3000/convocatoria/${id}/fecha-fin`, { fechaFin })
-        .then(function (response) {
-            console.log(response);
-        })
-};
-
-export const deleteConvocatoria = async (id: string): Promise<void> => {
-    try {
-        await axios.delete(`http://localhost:3000/convocatoria/${id}`);
-        console.log("Convocatoria eliminada correctamente");
-    } catch (error) {
-        console.error("Error al eliminar la convocatoria", error);
-        throw error;
-    }
-};
 
 export const registrarUsuario = async (
     nombre: string,
